Add update_env input to copilot_deploy
Adding this input so that multiple jobs can use this action at one time.
The copilot env only really needs to be updated once, so adding this to be able to control which job should do that.

See [here](https://github.com/LumioHX/solflow-api/actions/runs/5094935140) for an example of the job failing w/o the input

[here](https://github.com/LumioHX/solflow-api/actions/runs/5334165987) of it passing with the input

An example of how this is used in
[solflow-api](https://github.com/LumioHX/solflow-api/blob/master/.github/workflows/shared-deploy-2.yml#L110) currently: